+++ src/CurseWhenFriendDies/Plugin.cs
@@ -1,4 +1,5 @@
 ﻿using BepInEx;
+using BepInEx.Configuration;
 using BepInEx.Logging;
 using HarmonyLib;
 
@@ -9,11 +10,16 @@ public partial class Plugin : BaseUnityPlugin
 {
     internal static ManualLogSource Log { get; private set; } = null!;
 
+    internal static ConfigEntry<int> curseAmount;
+    internal static ConfigEntry<int> reviveRewardAmount;
+
     private void Awake()
     {
         Log = Logger;
         Log.LogInfo($"Plugin CurseWhenFriendDies is loaded!");
         new Harmony("com.github.DaxxPurpura.CurseWhenFriendDies").PatchAll();
 
+        curseAmount = Config.Bind("General", "Curse Amount", 5, "Amount of curse applied per death (1 unit = 2.5).");
+        reviveRewardAmount = Config.Bind("General", "Revive Reward Amount", 5, "Amount of curse removed per revive (1 unit = 2.5).");
     }
 }