+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.0] - 2025-08-20
+
+- Added configurable curse units.
+- Added revival reward.
+
 ## [1.0.0] - 2025-08-01
 
 - Initial release.